// backup/features/account/types/types.ts
export interface Account {
    id: number
    created_at: string | Date
    updated_at: string | Date
    user_id: number
    type: string
    suscription_id: string | null
}

export interface Store {
    id: number
    name: string
    description: string | null
    logo: string | null
    slogan: string | null
    created_at: Date
    updated_at: Date
    user_id: number
    slug: string
    subdomain: string
}

export interface UserType {
    id: number
    email: string
    avatar: string | null
    banner: string | null
    created_at: string | Date
    first_name: string | null
    last_name: string | null
    username: string | null
    phone?: string | null
    password?: string
    updated_at?: string | Date
    supabase_user_id: string | null
    Account?: Account[]
    Store?: Store[]
}